[fix] 즐겨찾기과 알람 controller 분리
